requirements.txt
pydub
wave
pyaudio
tk
pyfluidsynth